2 'Main' Changes:
1. Fixed a RIDICULOUS bit of incestous code - which was that Get-ConnectionObject ... set up a connection, then called into Test-Connection to ... test (good idea).
Only, Test-Connection was getting the COMMAND OBJECT to use from ... Get-CommandObject.
I've banged my head on this a gob of times in the past - cursing problems with "Get Connection" doing command-y stuff - when ... Get-CommandObject comes LATER in the pipeline.
So. Yeah. Anyhow.
Fixed that. (Tracking this crap down took a solid 20-30 minutes).
2. Finally, now, binding command-timeout so that it's worth a damn. (took all of 5 minutes (tops)) to add + test.